% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_tester.R
\name{add_bias}
\alias{add_bias}
\title{add_bias}
\usage{
add_bias(series, mu_bias, sigma_bias, mult_change)
}
\arguments{
\item{series}{A column vector with the original series}

\item{mu_bias}{The relative or absolute bias to be added in the mean}

\item{sigma_bias}{The relative or absolute bias to be added in the standard deviation}

\item{mult_change}{(Optional) A flag that indicates if projected changes should be computed as multiplicative (fut = hist*delta) or  additive (fut = hist + delta) changes. mult_change = 1 or True: Multiplicative (default) ; mult_change = 0 or False: Additive}
}
\value{
An array with the biased series
}
\description{
Adds bias based on a equi-probability transform
}
\examples{
add_bias(eries, mu_bias, sigma_bias, mult_change)
}
