% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climQMBC2-package.R
\docType{package}
\name{climQMBC2-package}
\alias{climQMBC2}
\alias{climQMBC2-package}
\title{climQMBC2: Climatic variables Quantile Mapping Bias Correction}
\description{
This package contains five bias correction methods: 1) Quantile Mapping (QM), 2) Detrended Quantile Mapping (DQM), 3) Quantile Delta Mapping (QDM), 4) Unbiased Quantile Mapping (UQM), and Scaled Distribution Mapping (SDM). Correction is performed to monthly or annual precipitation or temperature data in a single location.
}
\keyword{internal}
