% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{day_centered_moving_window}
\alias{day_centered_moving_window}
\title{day_centered_moving_window}
\usage{
day_centered_moving_window(series, day_win)
}
\arguments{
\item{day_win}{An integer indicating how many days to consider backwards and forward to get the statistics of each calendar day.The length of the window will be (2*win_day-1). For example, day_win=15 -> window of 29.}

\item{series_matrix}{A matrix of daily data, without considering leap days. The number of rows is 365 and number of columns is the  number of years of the series. [365, years]}
}
\value{
series_moving: A 3D array of daily data, without considering leap days, with a dimension that considers a centered moving window for each day of the year. [365, 2*win-1, years]
}
\description{
This function transform a 2D matrix of dimension [days in year, years] to a 3D array, where the new dimension is a centered moving window for each day of the year.
}
\examples{
day_centered_moving_window(series, day_win)
}
