% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatQM.R
\name{formatQM}
\alias{formatQM}
\title{formatQM}
\usage{
formatQM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)
}
\arguments{
\item{obs, mod}{A column vector of monthly or annual observed (modeled) data
(temperature or precipitation).

If monthly frequency is specified, the length of this vector is 12 times the
number of years \verb{[12 x ny]}.

If annual frequency is specified, the length of this vector is equal to the
number of years \verb{[ny]}.}

\item{var}{A flag that identifies if data are temperature or precipitation.
This flag tells the getDist function if it has to discard distribution
functions that allow negative numbers, and if the terms in the correction
equations are multiplied/divided or added/subtracted. Temperature: \code{var = 0};
Precipitation: \code{var = 1}}

\item{frq}{A string specifying if the input is annual or monthly data. If not
specified, it is set monthly as default. Monthly: \code{frq = 'M'}; Annual: \code{frq = 'A'}}

\item{pp_threshold}{A float indicating the threshold to consider physically
null precipitation values.}

\item{pp_factor}{A float indicating the maximum value of the random values
that replace physically null precipitation values.}
}
\value{
\code{list(obs, mod)}, \code{obs} and \code{mod} have the same structure:
\subsection{\code{data}:}{

A column vector of monthly or annual observed data (temperature or
precipitation). If monthly frequency is specified, the length of this
vector is 12 times the number of observed years \verb{[12, y_obs]}. If annual
frequency is specified, the length of this vector is equal to the number of
observed years \verb{[1, y_obs]}.
}

\subsection{\code{coef}:}{
\itemize{
\item \code{mu}: A column vector of mean values of the series. \verb{[12,1]} if the series
consider monthly data and \verb{[1,1]} if the series consider annual data.
\item \code{sigma}: A column vector of standard deviation of the series. \verb{[12,1]} if
the series consider monthly data and \verb{[1,1]} if the series consider annual
data.
\item \code{skew}: A column vector of skewness of the series. \verb{[12,1]} if the series
consider monthly data and \verb{[1,1]} if the series consider annual data.
\item \code{skewy}: A column vector of skewness of the logarithm of the series.
\verb{[12,1]} if the series consider monthly data and \verb{[1,1]} if the series
consider annual data.
}
}

\subsection{\code{nyear}: length of year}{
}
}
\description{
This function formats the inputs and gets basic statistics for the different
Quantile Mapping (QM, DQM, QDM, UQM and SDM) methods available in the
climQMBC package. If monthly data is specified, the input series will be
reshaped to a matrix of 12 rows and several columns equal to the number of
years of each series. If annual data is specified, the input is reshaped to a
row vector with same entries as the input series. For precipitation,
physically null values (values below \code{pp_threshold}) are replaced by random
positive values below pp_factor.
}
