% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pp_threshold_mod}
\alias{get_pp_threshold_mod}
\title{get_pp_threshold_mod}
\usage{
get_pp_threshold_mod(obs, mod, pp_threshold)
}
\arguments{
\item{obs}{A column vector of daily observed data, without  considering leap  days. The length of the column vector  should by a multiple of 365. [ndata_obs, 1]}

\item{mod}{A column vector of daily modeled data, without  considering leap  days. The length of the column vector should by a multiple of 365. [ndata_mod, 1]}

\item{pp_threshold}{A float indicating the threshold to consider no-rain values in the observed data.}
}
\value{
pp_threshold_mod: A float indicating the threshold to consider no-rain values in the modeled data.
}
\description{
This functions gets a no-rain value threshold for the modeled series by mathcing the no-rain days of the modeled series in the historical period with the observed series.
}
\examples{
get_pp_threshold_mod(obs, mod, pp_threshold)
}
