% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_tester.R
\name{get_synthetic_gamma}
\alias{get_synthetic_gamma}
\title{get_synthetic_gamma}
\usage{
get_synthetic_gamma(mu, sigma, yn)
}
\arguments{
\item{mu}{Mean of the series}

\item{sigma}{Standard deviation of the series}

\item{yn}{Number of years of the sample}
}
\value{
An array with the sampled values
}
\description{
Gets a random set of yn values of mean mu and standard deviation sigma based on a gamma distribution
}
\examples{
get_synthetic_gamma(mu, sigma, yn)
}
