% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_tester.R
\name{get_tester_stats}
\alias{get_tester_stats}
\title{get_tester_stats}
\usage{
get_tester_stats(bc_series, obj_series, mult_change)
}
\arguments{
\item{bc_series}{A column vector with the biased series}

\item{obj_series}{A column vector with the objective series}

\item{mult_change}{(Optional) A flag that indicates if projected changes should be computed as multiplicative (fut = hist*delta) or  additive (fut = hist + delta) changes. mult_change = 1 or True: Multiplicative (default) ; mult_change = 0 or False: Additive}
}
\value{
A list with the tester statistics
}
\description{
This function computes the tester statistics, including the NSE, KGE and ratios (or difference) in the mean, std. dev., skewness and several percentiles.
}
\examples{
get_tester_stats(bc_series, obj_series, mult_change)
}
