% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDist.R
\name{getCDFinv}
\alias{getCDFinv}
\title{Get the value associated to a certain probability}
\usage{
getCDFinv(PDF, Taot, mu, sigma, skew, skewy)
}
\arguments{
\item{PDF}{A column vector with an ID for the resulting distribution from the
KS test. \verb{[12,1]} if the series consider monthly data and \verb{[1,1]} if the
series consider annual data. The ID is related to the numeration of the
distribution listed in the description of this function.}

\item{Taot}{A column vector with the non-exceedance probability for each row
of the input series. \verb{[12,1]} if the series consider monthly data and \verb{[1,1]}
if the series consider annual data.}
}
\value{
\code{xhat}: A column vector with the values obtained when the inverse
cumulative distribution function is applied. \verb{[12,1]} if the series consider
monthly data and \verb{[1,1]} if the series consider annual data.
}
\description{
Evaluates the probability, Taot, in the respective inverse cumulative
distribution function assigned by the Kolmogorov-Smirnov (KS) test in the
\link{getDist}.
}
\section{Distributions}{
 The available distributions are:
\enumerate{
\item Normal distribution;
\item Log-Normal distribution;
\item Gamma 2 parameters distribution;
\item Gamma 3 parameters distribution (Pearson 3 parameters distribution);
\item Log-Gamma 3 parameters distribution (Log-Pearson 3 parameters
distribution);
\item Gumbel distribution;
\item Exponential distribution
}
\itemize{
\item For precipitation, only 2), 3) and 5) are considered (1, 4, 6, and 7 are
discarded).
\item For series with negative values, only 1), 3), 4), 6), and 7) are considered
(2, 3 and 5 are discarded).
}
}

\examples{
# temperature data
input <- climMOD
l <- with(input, formatQM(obs, mod, var))
print(str(l))

l_obs <- l$obs
l_mod <- l$mod

DIST_mod <- with(l_mod, getDist(data, coef, var))

cdf_mod <- with(l_mod, getCDF(DIST_mod, data, coef))

DIST_obs <- with(l_obs, getDist(data, coef, var))
QM_series <- getCDFinv(DIST_obs, cdf_mod, l_obs$coef)

r <- list(dist = DIST_mod, cdf = cdf_mod)
print(str(r))
}
