% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getStats}
\alias{getStats}
\title{getStats}
\usage{
getStats(series, frq)
}
\arguments{
\item{series}{An array of daily, monthly or annual data, without considering leap days. Possible input dimensions: 2D array: [sub-periods, years] or [sub-periods + days window, years] ; 3D array: [sub_periods, projected periods window, years] or [sub-periods + days window, projected periods window, years]}
}
\value{
mu: Mean values of each sub-period (and projected period if input is a 3D array).

sigma: Standar deviation values of each sub-period (and projected  period if input is a 3D array).

skew: Skewness values of each sub-period (and projected period if input is a 3D array).

skewy: Skewness values of the logarithm of theseries of each sub-period (and projected period if input is a 3D array).
}
\description{
This function computes the mean, standard deviation, skewness and skewness of the logarithmic values, for each sub-period within the year, according to the frequency initially defined. Statistics are computed along axis 1. NOTE: If input series is a 2D array, outputs will be a column vector. If input series is a 3D array, outputs will be a 2D array.
}
\examples{
getStats(series, frq)
}
