% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_DQM.R, R/map_QDM.R, R/map_QM.R,
%   R/map_SDM.R, R/map_UQM.R
\name{map_DQM}
\alias{map_DQM}
\alias{map_QDM}
\alias{map_QM}
\alias{map_SDM}
\alias{map_UQM}
\title{Quantile Mapping functions}
\usage{
map_DQM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)

map_QDM(
  obs,
  mod,
  var,
  frq = "M",
  pp_threshold = 1,
  pp_factor = 0.01,
  rel_change_th = 2,
  inv_mod_th = pp_threshold
)

map_QM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)

map_SDM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)

map_UQM(obs, mod, var, frq = "M", pp_threshold = 1, pp_factor = 0.01)
}
\arguments{
\item{obs, mod}{A column vector of monthly or annual observed (modeled) data
(temperature or precipitation).

If monthly frequency is specified, the length of this vector is 12 times the
number of years \verb{[12 x ny]}.

If annual frequency is specified, the length of this vector is equal to the
number of years \verb{[ny]}.}

\item{var}{A flag that identifies if data are temperature or precipitation.
This flag tells the getDist function if it has to discard distribution
functions that allow negative numbers, and if the terms in the correction
equations are multiplied/divided or added/subtracted.
\itemize{
\item \code{Temperature}  : \code{var = 0};
\item \code{Precipitation}: \code{var = 1}.
}}

\item{frq}{A string specifying if the input is annual or monthly data. If not
specified, it is set monthly as default. Monthly: \code{frq = 'M'}; Annual: \code{frq = 'A'}}

\item{pp_threshold}{A float indicating the threshold to consider physically
null precipitation values.}

\item{pp_factor}{A float indicating the maximum value of the random values
that replace physically null precipitation values.}

\item{rel_change_th}{(Optional) A float indicating the maximum scaling factor
(Eq. 4 of Cannon et al. (2015)) when the denominator is below inv_mod_th.}

\item{inv_mod_th}{(Optional) A float indicating the upper threshold of the
denominator of the scaling factor (Eq. 4 of Cannon et al. (2015)) to truncate
the scaling factor. This parameter is defined as default as the pp_threshold
parameter, described above.}
}
\value{
A column vector of monthly or annual modeled data (temperature or
precipitation) corrected by the QM method.
\itemize{
\item If monthly frequency is specified, the length of this vector is 12 times
the number of observed years \verb{[12 x nyear_mod, 1]}.
\item If annual frequency is specified, the length of this vector is the number
of observed  years \verb{[nyear_mod, 1]}.
}
}
\description{
\itemize{
\item Detrended Quantile Mapping (DQM): Cannon et al. (2015).
}

\itemize{
\item Quantile Delta Mapping (QDM): Cannon et al. (2015).
}

\itemize{
\item Quantile Mapping (QM): Cannon et al. (2015).
}

\itemize{
\item Scaled Distribution Mapping (SDM):

Proposed by Switanek et al. (2017). The historical period of the modeled
series is bias corrected as a scenario where the complete series is replaced
by the bias corrected series. On the other hand, for each year after the
historical period the method considers a projected period as a window whose
length is equal to the number of years of the historical period and ends in
the analyzed year. From the bias corrected series, only the last year is
saved and all the others are discarded. This approach aims to build a
continuum for the bias corrected series.
}

\itemize{
\item Unbiased Quantile Mapping (UQM): Chadwick et al. (2021).
}
}
\details{
Correction is performed to monthly or annual precipitation or
temperature data in a single location. An independent probability
distribution function is assigned to each month and to each projected period
based on the Kolmogorov-Smirnov test.

\emph{For DQM, QDM, and UQM:}

If annual frequency is specified, this is applied to the complete period.
Each projected period considers a window whose length is equal to the number
of years of the historical period and ends in the analyzed year.

Correction of the historical period is performed by the Quantile Mapping (QM)
method, as described by Cannon et al. (2015).
}
\references{
1.Cannon, A. J., S. R. Sobie, and T. Q. Murdock, 2015: Bias correction of GCM
precipitation by quantile mapping: How well do methods preserve changes in
quantiles and extremes? J. Climate, 28(17), 6938-6959,
https://doi.org/10.1175/JCLI-D-14-00754.1

3.Switanek, B. M., Troch, P., Castro, L. C., Leuprecht, A., Chang, H. I.,
Mukherjee, R., and Demaria, M. C. E. (2017) Scaled distribution mapping: A
bias correction method that preserves raw climate model projected changes.
Hydrology & Earth System Sciences, 21, 2649-2666,
https://doi.org/10.5194/hess-21-2649-2017.

2.Chadwick et al. (2022) \verb{[under revision]}
}
