% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{projected_backward_moving_window}
\alias{projected_backward_moving_window}
\title{projected_backward_moving_window}
\usage{
projected_backward_moving_window(series, projected_win, frq)
}
\arguments{
\item{series}{A 2D or 3D array of daily, monthly or annual observed data, without considering leap days. For daily, monthly and annual frequency, the number of rows will be 365, 12 and 1, respectively. If daily data, it is a 3D array that has a  dimension with a centered moving window for each day of the year. [1, 12 or 365, years] or [365, day centered window, years]}

\item{projected_win}{An integer the length of the backward moving window.}

\item{frq}{A string specifying if the input frequency is daily, monthly or annual. Daily:     frq = 'D' ; Monthly:   frq = 'M' ; Annual:    frq = 'A'}
}
\value{
win_series: An array of daily, monthly or annual future data, without considering leap days, with a dimension associated to a backward moving window for each projected period. Possible dimensions: 3D array: [sub-periods, projected periods window, years] ; 4D array: [sub-periods, days window, projected periods window, years]
}
\description{
This function transform a 2D or 3D array of dimensions [days in year, year] or [days in year, centered window for each day, years] to a 3D or 4D array adding a new dimension to represent a backward moving window for each  projected period.
}
\examples{
projected_backward_moving_window(series, projected_win, frq)
}
