% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{climQMBC example report}
\usage{
report(obs, mod, var, fun, y_init, y_wind)
}
\arguments{
\item{obs, mod}{A column vector of monthly or annual observed (modeled) data
(temperature or precipitation).

If monthly frequency is specified, the length of this vector is 12 times the
number of years \verb{[12 x ny]}.

If annual frequency is specified, the length of this vector is equal to the
number of years \verb{[ny]}.}

\item{var}{A flag that identifies if data are temperature or precipitation.
This flag tells the getDist function if it has to discard distribution
functions that allow negative numbers, and if the terms in the correction
equations are multiplied/divided or added/subtracted. Temperature: \code{var = 0};
Precipitation: \code{var = 1}}

\item{fun}{(Optional) A list of strings with the desired bias correction
methods to be reported. If this input is not recieved by the function, all
available bias correction methods will be reported.

The methods supported are:
\itemize{
\item a) 'QM' : Quantile Mapping;
\item b) 'DQM': Detrended Quantile Mapping;
\item c) 'QDM': Quantile Delta Mapping;
\item d) 'UQM': Unbiased Quantile Mapping;
\item e) \code{SDM}: Scaled Distribution Mapping
}

Default: fun = \verb{['QM','DQM','QDM','UQM','SDM']}}

\item{y_init}{(Optional) First year of the observed and modeled series
(integer). Default: y_init = 0}

\item{y_wind}{(Optional) A list of integers with the year of the center of
the projected periods to be reported. Default: y_wind = \verb{[int(y_obs+y_obs/2), int(y_mod-y_obs/2)]} This value sets a first projected period just after the
end of historical period, and a second projected period just before the end
of the modeled series.}
}
\value{
A list with the five methods implemented in the climQMBC ppackage:
QM, DQM, QDM, UQM, and SDM.
}
\description{
This function generates two report of the performance of the different
methods (QM, DQM, QDM, UQM and SDM) available in the climQMBC package. These
reports are based on the mean and standard deviation of the series in the
historical and future projected periods.
}
\details{
\itemize{
\item \verb{a summary table}: with the overall performance of the QM method in the
historical period and of the different methods in future projected periods.
The methods performance is addressed by comparing its variations in the mean
and standard deviation with the ones of the modeled data.
\item \verb{three figures}:
\itemize{
\item Figure 1 shows the cumulative distribution of the observed, modeled, and
corrected series in the historical and future period. This figure also shows
the complete time series.
\item Figure 2 and 3 shows the monthly mean and standard deviation,
respectively, of each series in the historical and future projected periods.
In these two figures, in the future projected periods, the observed series is
replaced by an objective series which is computed as the observed series
(monthly mean or standard deviation), scaled by the variation between the
projected and the historical period of the modeled series. Each projected
period is centered in a moving window whose length is equal to the length of
the historical period.
}
}
}
\examples{
report(obs, mod, var) 
report(obs, mod, var, fun=c('QDM','UQM','SDM'),y_init = 1979,y_wind = c(2035,2060,2080))

}
