% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_norain_to_nan}
\alias{set_norain_to_nan}
\title{set_norain_to_nan}
\usage{
set_norain_to_nan(series_moving, pp_threshold, pp_factor, min_rainday)
}
\arguments{
\item{series_moving}{A 2D or 3D array of daily, monthly or annual observed data, without considering leap days. For daily, monthly and annual frequency, the number of rows will be 365, 12 and 1, respectively. If daily data, it is a 3D array that has a  dimension with a centered moving window for each day of the year. [1, 12 or 365, years] or [365, day centered window, years]}

\item{pp_threshold}{A float indicating the threshold to consider no-rain values.}

\item{pp_factor}{A float which multiplied to pp_threshold indicates the maximum value of the random values that replace no-rain values.}

\item{min_rainday}{(Optional) Minimum amount of values to keep for each sub-period and projected period, to ensure a minimum amount to fit a  distribution. Default is 30}
}
\value{
series_moving: The input, but with no-rain values replaced with nans.
}
\description{
This function replace no-rain values with nans, leaving a minimum amout of values (min_rainday) to fit a distribution. If fewer than min_rainday values are above pp_threshold, random small values will be added.
}
\examples{
set_norain_to_nan(series_moving, pp_threshold, pp_factor)
set_norain_to_nan(series_moving, pp_threshold, pp_factor, min_rainday=20)
}
