% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_tester.R
\name{synthetic_tester}
\alias{synthetic_tester}
\title{synthetic_tester}
\usage{
synthetic_tester(obs, mod, allow_negatives, mult_change, SDM_var, yn)
}
\arguments{
\item{obs}{A column vector of monthly observed data.}

\item{mod}{A column vector of monthly modeled or GCM data.}

\item{allow_negatives}{A flag that identifies if data allows negative values and also to replace no-rain values with random small  values (Chadwick et al., 2023) to avoid numerical problems with the probability distribution functions. allow_negatives = 1 or True: Allow negatives (default) ; allow_negatives = 0 or False: Do not allow negative}

\item{mult_change}{A flag that indicates if projected changes should be computed as multiplicative (fut = hist*delta) or additive (fut = hist + delta) changes. mult_change = 1 or True: Multiplicative (default) ; mult_change = 0 or False: Additive}

\item{SDM_var}{A flag that identifies if data are temperature or precipitation. Temperature:   var = 0 ; Precipitation: var = 1}

\item{yn}{Number of years of the sample}
}
\value{
A table with the tester statistics
}
\description{
This function This function runs the synthetic tester function to compare the performance of different methods based on a synthetic example rather than the original data, to compare the theoretical performance.
}
\examples{
synthetic_tester(obs, mod, allow_negatives, mult_change, SDM_var, yn)
}
