% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatQM}
\alias{formatQM}
\title{formatQM}
\usage{
formatQM(series_, allow_negatives, frq, pp_threshold, pp_factor)
}
\arguments{
\item{series_}{A column vector of daily data, without considering leap days. The length of the column vector should by a multiple of 365. [ndata_obs, 1]}

\item{allow_negatives}{A flag that identifies if data allows negative values and also to replace no-rain values with random small values (Chadwick et al., 2023) to avoid numerical problems with the probability distribution functions. allow_negatives = 1 or True: Allow negatives ; allow_negatives = 0 or False: Do not allow negative}

\item{frq}{A string specifying if the input frequency is daily, monthly or annual. Daily:     frq = 'D' ; Monthly:   frq = 'M' ; Annual:    frq = 'A'}

\item{pp_threshold}{A float indicating the threshold to consider no-rain values.}

\item{pp_factor}{A float which multiplied to pp_threshold indicates the maximum value of the random values that replace no-rain values.}
}
\value{
years: Number of years of the series.

series_matrix: A matrix of daily, monthly or annual observed data, without considering leap days. For daily, monthly and annual frequency, the number of rows will be 365, 12 and 1, respectively. The number of columns will be equal to the number of years. [1, 12 or 365, years]
}
\description{
This function formats a time series from a column vector to a matrix. The number of rows depends on the frequency of the data (1, 12 and 365 for annual, monthyl and daily data, respectively). If negative values are not allowed, determined by allow_negatives=1, no-rain values (values below pp_threshold) are replaced by random positive values below the product pp_threshold*pp_factor.
}
\examples{
formatQM(series_, allow_negatives, frq, pp_threshold, pp_factor)
}
