% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCDF}
\alias{getCDF}
\title{getCDF}
\usage{
getCDF(pdf, series, mu, sigma, skew, skewy)
}
\arguments{
\item{pdf}{A vector with an ID for the resulting distribution from the KS test. The ID is related to  the numeration of the distribution listed in the description of this function.}

\item{series}{An array of daily, monthly or annual data, without considering leap days. Possible input dimensions: 2D array: [sub-periods, years] ; [sub-periods + days window, years]}

\item{mu}{A vector with mean values of each sub-period.}

\item{sigma}{A vector with standard deviation values of each sub-period.}

\item{skew}{A vector with skewness values of each sub-period.}

\item{skewy}{A vector with skewness values of the logarithm of the series of each sub-period.}
}
\value{
prob: A matrix with the non-exceedance probability for value row of the input series.
}
\description{
This function evaluates each row of the series in the respective cumulative distribution function assigned by the Kolmogorov-Smirnov (KS) test in the getDist function of the climQMBC package.
}
\details{
The available distributions are: 1) Normal distribution; 2) Log-Normal distribution; 3) Gamma 2 parameters distribution; 4) Gamma 3 parameters distribution (Pearson 3 parameters distribution); 5) Log-Gamma 3 parameters distribution (Log-Pearson 3 parameters distribution); 6) Gumbel distribution; 7) Exponential distribution
}
\examples{
getCDF(pdf,series,mu,sigma,skew,skewy)
}
