% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCDFinv}
\alias{getCDFinv}
\title{getCDFinv}
\usage{
getCDFinv(pdf, prob, mu, sigma, skew, skewy)
}
\arguments{
\item{pdf}{A vector with an ID for the resulting distribution from the KS test. The ID is related to  the numeration of the distribution listed in the description of this function.}

\item{prob}{A matrix with the non-exceedance probability for value row of the input series.}

\item{mu}{A vector with mean values of each sub-period.}

\item{sigma}{A vector with standard deviation values of each sub-period.}

\item{skew}{A vector with skewness values of each sub-period.}

\item{skewy}{A vector with skewness values of the logarithm of the series of each sub-period.}
}
\value{
xhat: A matrix with the values obtained when the inverse cumulative distribution function is applied to prob.
}
\description{
This function evaluates the probability, prob, in the respective inverse cumulative distribution function assigned by the Kolmogorov-Smirnov (KS) test in the getDist function of the climQMBC package.
}
\details{
The available distributions are: 1) Normal distribution; 2) Log-Normal distribution; 3) Gamma 2 parameters distribution; 4) Gamma 3 parameters distribution (Pearson 3 parameters distribution); 5) Log-Gamma 3 parameters distribution (Log-Pearson 3 parameters distribution); 6) Gumbel distribution; 7) Exponential distribution
}
\examples{
getCDFinv(PDF,Taot,mu,sigma,skew,skewy)
}
