% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDist}
\alias{getDist}
\title{getDist}
\usage{
getDist(series, allow_negatives, mu, sigma, skew, skewy)
}
\arguments{
\item{series}{An array of daily, monthly or annual data, without considering leap days. Possible input dimensions: 2D array: [sub-periods, years] or [sub-periods + days window, years]}

\item{allow_negatives}{A flag that identifies if data allows negative values and also to replace no-rain values with random small values (Chadwick et al., 2023) to avoid numerical problems with the probability distribution functions. allow_negatives = 1 or True: Allow negatives ; allow_negatives = 0 or False: Do not allow negative}

\item{mu}{A vector with mean values of each sub-period.}

\item{sigma}{A vector with standard deviation values of each sub-period.}

\item{skew}{A vector with skewness values of each sub-period.}

\item{skewy}{A vector with skewness values of the logarithm of the series of each sub-period.}
}
\value{
pdf: A vector with an ID for the resulting distribution from the KS test. The ID is related to  the numeration of the distribution listed in the description of this function. This ID is used in the getCDF and getCDFinv  functions of the climQMBC package.
}
\description{
This function assigns an independent probability distribution function to each row of the input series by comparing the empirical probability distribution function with seven distributions based on the Kolmogorov-Smirnov (KS) test.
}
\details{
The available distributions are: 1) Normal ; 2) Log-Normal ; 3) Gamma 2 parameters ; 4) Gamma 3 parameters ; (Pearson 3 parameters) ; 5) Log-Gamma 3 parameters ; 6) Gumbel ; 7) Exponential

For allow_negatives=0, only 2), 3) and 5) are considered (1, 4, 6, and 7 are discarded). For series with negative values, only 1), 3), 4), 6), and 7) are considered (2, 3 and 5 are discarded).
}
\examples{
getDist(series, allow_negatives, mu, sigma, skew, skewy)
}
