% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{QM}
\alias{QM}
\title{Quantile Mapping}
\usage{
QM(
  obs,
  mod,
  allow_negatives,
  frq,
  pp_threshold,
  pp_factor,
  day_win,
  user_pdf,
  pdf_obs,
  pdf_mod
)
}
\arguments{
\item{obs}{A column vector of daily, monthly or annual observed data, without considering leap days. If daily frequency is specified, the length of the column vector should by a multiple of 365 and for monthly frequency, it should be a multiple of 12. [ndata_obs, 1]}

\item{mod}{A column vector of daily, monthly or annual modeled or GCM data, without considering leap days. If daily frequency is specified, the length of the column vector should by a multiple of 365 and for monthly frequency, it should be a multiple of 12. [ndata_mod, 1]}

\item{allow_negatives}{(Optional) A flag that identifies if data allows negative values and also to replace no-rain values with random small values (Chadwick et al., 2023) to avoid numerical problems with the probability distribution functions. allow_negatives = 1 or True: Allow negatives (default) ; allow_negatives = 0 or False: Do not allow negative}

\item{frq}{(Optional) A string specifying if the input frequency is daily, monthly or annual. Daily:     frq = 'D' ; Monthly:   frq = 'M' ; Annual:    frq = 'A' (default)}

\item{pp_threshold}{(Optional) A float indicating the threshold to consider no-rain values. Default is 1.}

\item{pp_factor}{(Optional) A float which multiplied to pp_threshold indicates the maximum value of the random values that replace no-rain values. Default is 1/100.}

\item{day_win}{(Optional) (only for frq='D') An integer indicating how many days to consider backwards and forward to get the statistics of each calendar day.The length of the window will be  (2*win_day-1). For example, day_win=15 -> window of 29. Default: win = 1}

\item{user_pdf}{(Optional) A flag indicating if the user will define the probability distribution functions (pdf) for the observed and modeled series. The distributions will be the same for all periods and sub-periods. user_pdf = 1 or True: User defines the pdf ; user_pdf = 0 or False: pdf defined by the Kolmogorov -Smirnov test (default)}

\item{pdf_obs}{(Optional) An integer indicating the probability distribution function (pdf) to be used for the observed data. The pdf will be the same for all periods and sub-periods. Default: None}

\item{pdf_mod}{(Optional) An integer indicating the probability distribution function (pdf) to be used for the modeled data. The pdf will be the same for all periods and sub-periods. Default: None}
}
\value{
QM_series: A column vector of data bias corrected with the QM method. [ndata_mod, 1]
}
\description{
This function performs bias correction of modeled series based on observed data by the Quantile Mapping (QM) method, as described by Cannon et al. (2015). Correction is performed to daily, monthly or annual data in a single location. An independent probability distribution function is assigned to  each sub-annual period based on the Kolmogorov-Smirnov test.
}
\details{
The available distributions are: 1) Normal ; 2) Log-Normal ; 3) Gamma 2 parameters ; 4) Gamma 3 parameters ; (Pearson 3 parameters) ; 5) Log-Gamma 3 parameters ; 6) Gumbel ; 7) Exponential

NOTE: This routine considers that obs and mod series start in the same day/month/year and are continuous until the end day/month/year.

Cannon, A. J., S. R. Sobie, and T. Q. Murdock, 2015: Bias correction of GCM precipitation by quantile mapping: How well do methods preserve changes in quantiles and extremes? J. Climate, 28(17), 6938-6959, https://doi.org/10.1175/JCLI-D-14-00754.1
}
\examples{
QM(obs,mod)
QM(obs,mod,frq='A')
QM(obs,mod,frq='M')
QM(obs,mod,frq='D')
QM(obs,mod,allow_negatives=1)
QM(obs,mod,allow_negatives=0)
QM(obs,mod,allow_negatives=0,frq='D',pp_factor=1/10000,day_win=15)
}
