% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{climQMBC example report}
\usage{
report(
  obs,
  mod,
  SDM_var,
  mult_change,
  allow_negatives,
  fun,
  y_init,
  y_wind,
  user_pdf,
  pdf_obs,
  pdf_mod
)
}
\arguments{
\item{obs}{A column vector of monthly observed data. The length of the column vector should by a multiple of 12. [ndata_obs, 1]}

\item{mod}{A column vector of monthly modeled or GCM data.The length of the column vector should by a multiple of 12. [ndata_obs, 1]}

\item{mult_change}{(Optional) A flag that indicates if projected changes should be computed as multiplicative (fut = hist*delta) or additive (fut = hist + delta) changes. mult_change = 1 or True: Multiplicative (default) ; mult_change = 0 or False: Additive}

\item{allow_negatives}{(Optional) A flag that identifies if data allows negative values and also to replace no-rain values with random small  values (Chadwick et al., 2023) to avoid numerical problems with the probability distribution functions. allow_negatives = 1 or True: Allow negatives (default) ; allow_negatives = 0 or False: Do not allow negative}

\item{fun}{(Optional) A list of strings with the desired bias correction methods to be reported. If this input is not recieved by the function, all bias correction methods available in the climQMBC package will be reported. The methods supported are: a) 'QM' : Quantile Mapping; b) 'DQM': Detrended Quantile Mapping; c) 'QDM': Quantile Delta Mapping; d) 'UQM': Unbiased Quantile Mapping; Default: fun = ['QM','DQM','QDM','UQM','SDM']}

\item{y_init}{(Optional) First year of the observed and modeled series (integer). Default: y_init = 0}

\item{y_wind}{(Optional) A list of integers with the year of the center of the projected periods to be reported. Default: y_wind = [int(y_obs+y_obs/2), int(y_mod-y_obs/2)] This value sets a first projected period just after the end of historical period, and a second projected period just before the end of the modeled series.}

\item{user_pdf}{(Optional) A flag indicating if the user will define the probability distribution functions (pdf) for the observed and modeled series. The distributions will be the same for all periods and sub-periods. user_pdf = 1 or True: User defines the pdf ; user_pdf = 0 or False: pdf defined by the Kolmogorov -Smirnov test (default). NOTE: The available distributions for pdf_obs and pdf_mod are:  1) Normal ; 2) Log-Normal ; 3) Gamma 2 parameters ; 4) Gamma 3 parameters ; 5) Log-Gamma 3 parameters ; 6) Gumbel ; 7) Exponential}

\item{pdf_obs}{(Optional) An integer indicating the probability distribution function (pdf) to be used for the observed data. The pdf will be the same for all periods and sub-periods. Default: FALSE}

\item{pdf_mod}{(Optional) An integer indicating the probability distribution  function (pdf) to be used for the modeled data. The pdf will be the same for all periods and sub-periods. Default: FALSE}

\item{var}{A flag that identifies if data are temperature or precipitation. Temperature:   var = 0 ; Precipitation: var = 1}
}
\value{
A list with the five methods implemented in the climQMBC ppackage: QM, DQM, QDM, UQM, and SDM.
}
\description{
This function generates two report of the performance of the different methods (QM, DQM, QDM, UQM and SDM) available in the climQMBC package. These reports are based on the mean and standard deviation of the series in the historical and future projected periods.
}
\details{
The first report is a summary table with the overall performance of the QM method in the historical period and of the different methods in future projected periods. The methods performance is addressed by comparing its variations in the mean and standard deviation with the ones of the modeled data.

The second report consist of three figures. Figure 1 shows the cumulative distribution of the observed, modeled, and corrected series in the historical and future period. This figure also shows the complete time series. Figure 2 and 3 shows the monthly mean and standard deviation, respectively, of each series in the historical and future projected periods. In these two figures, in the future projected periods, the observed series is replaced by an objective series which is computed as the observed series (monthly mean or standard deviation), scaled by the variation between the projected and the historical period of the modeled series. Each projected period is centered in a moving window whose length is equal to the length of the historical period.

NOTE: This routine considers that obs and mod series start in the same day/month/year and are continuous until the end day/month/year.
}
\examples{
report(obs, mod, SDM_var)
report(obs, mod, SDM_var, mult_change, allow_negatives)
report(obs, mod, SDM_var,fun=['QDM','UQM','SDM'],y_init = 1979,y_wind = [2035,2060,2080])
}
